Refactor General Application Repo and corresponding BSON reader
